services/ingestion/requirements.txt
# Parsing & layout
docling==2.5.0
pytesseract==0.3.10
python-magic==0.4.27
Pillow==10.2.0
lxml==5.1.0

# NLP / text
tiktoken==0.6.0
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4

# ML / retrieval
openai==1.14.0
pinecone-client==3.2.2
rank-bm25==0.2.2
sentence-transformers==2.6.1

# Infrastructure
fastapi==0.110.0
uvicorn[standard]==0.28.0
celery==5.3.6
redis==5.0.1
boto3==1.34.0
httpx==0.27.0
tenacity==8.2.3
pydantic==2.6.3
pydantic-settings==2.2.1
structlog==24.1.0
prometheus-client==0.20.0
pytest==8.1.0
pytest-asyncio==0.23.5